filtro de restaurantes

+++ src/main/java/ar/edu/unlam/tallerweb1/controladores/RestaurantController.java
@@ -18,7 +18,7 @@ public class RestaurantController {
     private RestaurantService restaurantService;
 
     @RequestMapping(value = "/restaurants/{idRestaurant}", method = RequestMethod.GET)
-    public ModelAndView getRestaurant(@PathVariable("idRestaurant") Integer id){
+    public ModelAndView getRestaurant(@PathVariable("idRestaurant") Long id){
         ModelMap model = new ModelMap();
 
         try{
